+++ Examples/YandexMobileAdsExample/YandexMobileAdsExampleUITests/Tests/MobileMediationNativeTest.swift
@@ -22,7 +22,7 @@ final class MobileMediationNativeTest: BaseTest {
         mobileMediationPage.openNative()
         adPage.selectAdapter(adapter)
         adPage.tapLoadAd()
-        try assertAdLoaded(stateLabel: adPage.stateLabel)
+        guard assertAdLoaded(stateLabel: adPage.stateLabel) else { return }
         adPage.assertAdDisplayed()
     }
 }