+++ changelog/mediation/mobileads-mintegral/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 7.7.7.2
+
+### Updated
+
+- Updated minimum supported `YandexMobileAds` version to `7.14.0`
+
 ## Version 7.7.7.1
 
 ### Updated